wip - streamlined visibility checks
